requirements.txt
Pillow==10.3.0
qrcode==7.4.2
python-barcode==0.15.1